+++ kwcoco/formats/kpf.py
@@ -2,13 +2,53 @@
 WIP:
 
 Conversions to and from KPF format.
+
+
+.. code::
+
+    +------------+----------------------------+
+    | Packet     | Definition                 |
+    +------------+----------------------------+
+    | id         | object identifier          |
+    +------------+----------------------------+
+    | ts         | timestamp                  |
+    +------------+----------------------------+
+    | tsr        | timestamp range            |
+    +------------+----------------------------+
+    | loc        | location                   |
+    +------------+----------------------------+
+    | g          | bounding box               |
+    +------------+----------------------------+
+    | poly       | polygon                    |
+    +------------+----------------------------+
+    | conf       | confidence or likelihood   |
+    +------------+----------------------------+
+    | cset       | set of labels/ likelihoods |
+    +------------+----------------------------+
+    | act        | activity                   |
+    +------------+----------------------------+
+    | eval       | evaluation result          |
+    +------------+----------------------------+
+    | a          | attribute                  |
+    +------------+----------------------------+
+    | tag        | a packet / domain pair     |
+    +------------+----------------------------+
+    | kv         | key / value pair           |
+    +------------+-+--------------------------+
+
+
+References:
+    https://github.com/Kitware/DIVA/blob/master/doc/manuals/kpf.rst
 """
 
 
 def coco_to_kpf(coco_dset):
     """
-    import kwcoco
-    coco_dset = kwcoco.CocoDataset.demo('shapes8')
+    Example:
+        >>> from kwcoco.formats.kpf import *  # NOQA
+        >>> import kwcoco
+        >>> coco_dset = kwcoco.CocoDataset.demo('shapes8')
+        >>> coco_to_kpf(coco_dset)
     """
     import kwimage
     import ubelt as ub
@@ -39,6 +79,41 @@ def coco_to_kpf(coco_dset):
 
         packet = {'geom': geom}
         print('{}'.format(ub.urepr(packet, nl=0)))
+
+
+class KPFStream:
+    ...
+
+    @classmethod
+    def demo_packets(cls):
+        ...
+        import kwutil
+        import ubelt as ub
+        lines = ub.codeblock(
+            '''
+            { meta: "cmdline0: run_detector param1 param2..." }
+            { meta: "conf0: yolo person detector" }
+            { meta: "id0 domain: yolo person detector" }
+            { meta: "cmdline1: run_linker param1 param2..." }
+            { meta: "id1 domain: track linker hash 0x85913" }
+            { meta: "cmdline2: score_tracks param1 param2..." }
+            { meta: "overall track pd/fa count: 0.5 / 1" }
+            { meta: "eval0 domain against id0" }
+            { meta: "eval1 domain against id1" }
+            { meta: "id2 domain false negatives from official_ground_truth.kpf" }
+            { geom: { id0: 0, ts0: 101, g0: 515 419 525 430, conf0: 0.8, id1: 100, eval0: tp, eval1: tp } }
+            { geom: { id0: 1, ts0: 101, g0: 413 303 423 313, conf0: 0.3, id1: 102, eval0: fa, eval1: fa } }
+            { geom: { id0: 2, ts0: 102, g0: 517 421 527 432, conf0: 0.7, id1: 100, eval0: fa, eval1: tp } }
+            { geom: { id0: 3, ts0: 102, g0: 416 304 421 315, conf0: 0.2, id1: 102, eval0: fa, eval1: tp } }
+            { geom: { id2: 0, ts0: 101, g0: 600 550 605 610, eval0: fn, eval1: fn } }
+            { geom: { id2: 1, ts0: 102, g0: 603 553 608 615, eval0: fn, eval1: fn } }
+            ''').split('\n')
+        packets = [kwutil.Yaml.coerce(line) for line in lines if line]
+        return packets
+
+
+# class kpf_to_coco(packets):
+#     ...
 
 
 def demo():